+++ packages/iconoir-react-native/src/QrCode.tsx
@@ -1,0 +1,30 @@
+import * as React from 'react';
+import Svg, { SvgProps, Path } from 'react-native-svg';
+
+function SvgQrCode(
+  props: SvgProps,
+  svgRef?: React.Ref<React.Component<SvgProps>>
+) {
+  return (
+    <Svg
+      width="1.5em"
+      height="1.5em"
+      strokeWidth={1.5}
+      viewBox="0 0 24 24"
+      fill="none"
+      color="currentColor"
+      ref={svgRef}
+      {...props}
+    >
+      <Path
+        d="M15 8v7M12 3v3M18 9v6M9 18h12M18 21h3M6 12h3M6 6.011L6.01 6M12 12.011l.01-.011M3 12.011L3.01 12M12 9.011L12.01 9M12 15.011l.01-.011M15 21.011l.01-.011M12 21.011l.01-.011M21 12.011l.01-.011M21 15.011l.01-.011M18 6.011L18.01 6M9 3.6v4.8a.6.6 0 01-.6.6H3.6a.6.6 0 01-.6-.6V3.6a.6.6 0 01.6-.6h4.8a.6.6 0 01.6.6zM21 3.6v4.8a.6.6 0 01-.6.6h-4.8a.6.6 0 01-.6-.6V3.6a.6.6 0 01.6-.6h4.8a.6.6 0 01.6.6zM6 18.011L6.01 18M9 15.6v4.8a.6.6 0 01-.6.6H3.6a.6.6 0 01-.6-.6v-4.8a.6.6 0 01.6-.6h4.8a.6.6 0 01.6.6z"
+        stroke="currentColor"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </Svg>
+  );
+}
+
+const ForwardRef = React.forwardRef(SvgQrCode);
+export default ForwardRef;