+++ packages/iconoir-react-native/CHANGELOG.md
@@ -1,14 +1,21 @@
 # Changelog
-
 All notable changes to the [iconoir-react-native](https://www.npmjs.com/package/iconoir-react-native)
 npm package will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2021-07-05
+### Added
+- Support webpack tree shaking - Thanks @barnabasJ!
+
+### Changed
+- Updated `iconoir-react-native` to use iconoir `v4.4.0`.
+- Update dependencies.
+
 ## [2.0.1] - 2021-06-28
 ### Fixed
-- Don't bundle 'react-native-svg', fixes #54
+- Don't bundle 'react-native-svg', fixes #54.
 
 ## [1.0.0] - 2021-06-19
 ### Added